Implement very basic communication mechanisms of CastSocket
We can connect to a device, write messages to it, and log received
messages. We don't do anything with received messages (yet!) but this is
a good start. It was bizarrely tricky to even get here.

In particular: NSData is terribly unergonomic. How does it not have
primitives for writing byte values into it and reading them out of it
*at all*?