Stub out many more APIs to make Youtube almost usable
Youtube seems to use the chrome.cast APIs directly, but also only with
the Chrome User Agent for some reason. It seems like there *ought* to be
a way to get past that user agent check somehow, but for now we can get
the Cast icon to show up and be interactive when faking the user agent
via Safari's dev tools.

We will still need to stub the chrome.cast.Session class for this to
actually work...