Refactor so Youtube picks up our content scripts (#1)
* Try to patch the user agent to get YT to actually call our APIs

Works with a breakpoint set, but without it, remote.js seems to load
*before* our code, frustratingly...

* Load castable script sooner to improve YT compat

Also cleans up the compat code a bit. Don't recall why we seemed to need
to do the DOMContentLoaded dance before, but it seems fine without it
now.

* Clean up old content-loaded register-extension dance

I think the issue before was that we weren't falling back from
`document.head` to `document.documentElement` as a target for inserting
our content script, which is why it didn't get loaded. I picked up this
trick from Youtube's remote player script loading, funnily enough.