Make sockets more robust (#5)
* Try to improve robustness of CastSocket and reading from it

We seem to be getting disconnected by the chromecast for some reason,
and the existing error propagation isn't cutting it. This commit is the
first step to more reliably propagate network and other exceptions to
receivers so we can better handle errors (and perhaps, also, diagnose
them!)

* Track/log heartbeat timeout statistics

* Throw on failure to encode JSON messages instead of silently dropping

* Respond to PINGs correctly; move all heartbeats into Runner

* Cancel the Channel.receive channel when its input Channel cancels

It'd be really nice to have Kotlin's Flow here. We could perhaps create
an approximation based on what we're doing with CoChannel...