Eagerly manage Device state (#6)
* Move device management to its own class, with device lifecycles

The goal here is that *at least* when a device becomes "discovered" we
should interrogate its state to see if it's already casting something.
In particular, we don't want to lose the "active" device when we rebuild
the extension. That'll come in future commits

* Elect "active" device based on running applications

Note: we may actually want to go ahead and subscribe to the receiver
channel for known devices so we can detect apps launched outside of our
knowledge. On the other hand, Safari may destroy the ExtensionHandler
instance frequently enough that we'll do the query when necessary anyway

* Ensure initial devices get "managed"

The intention is to make sure our "active device" state is up-to-date
without having to maintain a connection to every device on the network

* Also inflate state.activeApp if appropriate