Stop implicitly passing Page to handlers by modifying the request
We should've done something like this a while ago, but now's as good a
time as any. In particular, this fixes Disney+ support, which we broke
while working on the events framework, since that introduced the
non-serializable value into the request.

Now that we *never* insert non-serializable values into the request
dictionary, and just pass it through as-is, things should be much easier
to reason about.